@@ -50,12 +50,11 @@
       </plugin>
       <!-- Maven Compiler Plugin -->
       <plugin>
-        <groupId>org.apache.maven.plugins</groupId>
         <artifactId>maven-compiler-plugin</artifactId>
-        <version>3.8.1</version>
+        <version>3.10.1</version>
         <configuration>
-          <source>15</source>
-          <target>15</target>
+          <source>17</source>
+          <target>17</target>
           <!-- avoid possible encoding issues -->
           <encoding>utf-8</encoding>
         </configuration>
@@ -64,7 +63,7 @@
       <plugin>
         <groupId>org.apache.maven.plugins</groupId>
         <artifactId>maven-jar-plugin</artifactId>
-        <version>3.2.0</version>
+        <version>3.3.0</version>
         <executions>
           <execution>
             <goals>
@@ -77,7 +76,7 @@
       <plugin>
         <groupId>org.apache.maven.plugins</groupId>
         <artifactId>maven-war-plugin</artifactId>
-        <version>3.3.1</version>
+        <version>3.3.2</version>
       </plugin>
       <!-- Exec Plugin for running npm -->
       <plugin>
@@ -104,7 +103,7 @@
       <plugin>
         <groupId>org.flywaydb</groupId>
         <artifactId>flyway-maven-plugin</artifactId>
-        <version>6.0.1</version>
+        <version>8.5.2</version>
         <configuration>
           <configFiles>
             <configFile>src/main/resources/flyway.conf</configFile>
@@ -132,7 +131,7 @@
     <commons-csv.version>1.8</commons-csv.version>
     <commons-fileupload.version>1.4</commons-fileupload.version>
     <c3p0.version>0.9.5.5</c3p0.version>
-    <eh-cache.version>2.10.6</eh-cache.version>
+    <ehcache.version>3.9.9</ehcache.version>
     <freemarker.version>2.3.31</freemarker.version>
     <guava.version>30.1.1-jre</guava.version>
     <guice.version>5.0.1</guice.version>
@@ -140,6 +139,7 @@
     <hierarchicalcontextrunner.version>4.12.2</hierarchicalcontextrunner.version>
     <jackson.version>2.13.2</jackson.version>
     <jackson-databind.version>2.13.2.2</jackson-databind.version>
+    <javax-cache.version>1.1.1</javax-cache.version>
     <jstl.version>1.2</jstl.version>
     <junit.version>4.13.2</junit.version>
     <mockito.version>3.9.0</mockito.version>
@@ -266,9 +266,15 @@
 
     <!-- EhCache -->
     <dependency>
-      <groupId>net.sf.ehcache</groupId>
+      <groupId>org.ehcache</groupId>
       <artifactId>ehcache</artifactId>
-      <version>${eh-cache.version}</version>
+      <version>${ehcache.version}</version>
+    </dependency>
+
+    <dependency>
+      <groupId>javax.cache</groupId>
+      <artifactId>cache-api</artifactId>
+      <version>${javax-cache.version}</version>
     </dependency>
 
     <!-- Jackson | Json/Xml serializers -->
