@@ -5,27 +5,22 @@ import gov.nysenate.ess.core.client.response.base.ListViewResponse;
 import gov.nysenate.ess.core.client.response.base.SimpleResponse;
 import gov.nysenate.ess.core.client.view.CacheStatsView;
 import gov.nysenate.ess.core.model.base.InvalidRequestParamEx;
-import gov.nysenate.ess.core.model.cache.CacheEvictEvent;
-import gov.nysenate.ess.core.model.cache.CacheEvictIdEvent;
-import gov.nysenate.ess.core.model.cache.CacheWarmEvent;
-import gov.nysenate.ess.core.model.cache.ContentCache;
-import gov.nysenate.ess.core.service.cache.EhCacheManageService;
+import gov.nysenate.ess.core.model.cache.CacheType;
+import gov.nysenate.ess.core.model.period.PayPeriodType;
+import gov.nysenate.ess.core.service.cache.EssCacheManager;
 import org.apache.shiro.authz.annotation.RequiresPermissions;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.Arrays;
 import java.util.EnumSet;
 import java.util.Set;
-import java.util.stream.Collectors;
 
 import static gov.nysenate.ess.core.controller.api.BaseRestApiCtrl.REST_PATH;
 
 @RestController
 @RequestMapping(value = REST_PATH + "admin/cache")
 public class CacheManagementCtrl extends BaseRestApiCtrl {
-
-    @Autowired EhCacheManageService cacheManageService;
-
     @Autowired EventBus eventBus;
 
     /**
@@ -42,11 +37,7 @@ public class CacheManagementCtrl extends BaseRestApiCtrl {
     @RequiresPermissions("admin:cache:get")
     @RequestMapping(value = "/stats", method = {RequestMethod.GET, RequestMethod.HEAD})
     public ListViewResponse<CacheStatsView> getCacheStats() {
-        return ListViewResponse.of(
-                cacheManageService.getCacheStatistics().stream()
-                        .map(CacheStatsView::new)
-                        .collect(Collectors.toList())
-        );
+        return ListViewResponse.of(Arrays.stream(CacheType.values()).map(CacheStatsView::new).toList());
     }
 
     /**
@@ -61,14 +52,14 @@ public class CacheManagementCtrl extends BaseRestApiCtrl {
      *
      * Path Variables:
      * @param cacheName String - the name of the cache to warm, use "all" to warm all caches
-     *                  @see ContentCache for cache names
+     *                  @see CacheType for cache names
      *
      * @return SimpleResponse - indicating cache warm success
      */
     @RequiresPermissions("admin:cache:put")
     @RequestMapping(value = "/{cacheName}", method = RequestMethod.PUT)
     public SimpleResponse warmCache(@PathVariable String cacheName) {
-        eventBus.post(new CacheWarmEvent(getAffectedCaches(cacheName)));
+        EssCacheManager.clearCaches(getAffectedCaches(cacheName), true);
         return new SimpleResponse(true, "warmed cache: " + cacheName, "cache-warm-success");
     }
 
@@ -82,14 +73,14 @@ public class CacheManagementCtrl extends BaseRestApiCtrl {
      *
      * Path Variables:
      * @param cacheName String - the name of the cache to evict, use "all" to evict all caches
-     *                  @see ContentCache for cache names
+     *                  @see CacheType for cache names
      *
      * @return SimpleResponse - indicating cache evict success
      */
     @RequiresPermissions("admin:cache:delete")
     @RequestMapping(value = "/{cacheName}", method = RequestMethod.DELETE)
     public SimpleResponse evictCache(@PathVariable String cacheName) {
-        eventBus.post(new CacheEvictEvent(getAffectedCaches(cacheName)));
+        EssCacheManager.clearCaches(getAffectedCaches(cacheName), false);
         return new SimpleResponse(true, "evicted cache: " + cacheName, "cache-evict-success");
     }
 
@@ -103,20 +94,19 @@ public class CacheManagementCtrl extends BaseRestApiCtrl {
      *
      * Path Variables:
      * @param cacheName String - the name of the cache to warm, use "all" to warm all caches
-     *                  @see ContentCache for cache names
+     *                  @see CacheType for cache names
      * Request Params:
      * @param key String - key for the element to evict
      *
      * @return SimpleResponse - indicating cache element evict success
      */
     @RequiresPermissions("admin:cache:delete")
-    @RequestMapping(value = "/{cacheName}", params = {"key"},method = RequestMethod.DELETE)
+    @RequestMapping(value = "/{cacheName}", params = {"key"}, method = RequestMethod.DELETE)
     public SimpleResponse evictCacheElement(@PathVariable String cacheName,
                                             @RequestParam String key) {
-        ContentCache cache = getEnumParameter("cacheName", cacheName, ContentCache.class);
-        Object parsedKey = parseElementKey(cache, key);
-        eventBus.post(new CacheEvictIdEvent<>(cache, parsedKey));
-
+        CacheType cache = getEnumParameter("cacheName", cacheName, CacheType.class);
+        Object parsedKey = cache.getTypedKey(key);
+        // TODO: evict content
         return new SimpleResponse(true,
                 "evicted element:" + key + " from cache:" + cacheName,
                 "cache-evict-success");
@@ -124,35 +114,12 @@ public class CacheManagementCtrl extends BaseRestApiCtrl {
 
     /** --- Internal Methods --- */
 
-    private Set<ContentCache> getAffectedCaches(String cacheName) {
+    private Set<CacheType> getAffectedCaches(String cacheName) {
         if ("all".equalsIgnoreCase(cacheName)) {
-            return ContentCache.getAllContentCaches();
+            return Set.of(CacheType.values());
         }
         return EnumSet.of(
-                getEnumParameter("cacheName", cacheName, ContentCache.class)
+                getEnumParameter("cacheName", cacheName, CacheType.class)
         );
     }
-
-    private Object parseElementKey(ContentCache cacheType, String key) {
-        Class<?> keyType = cacheType.getKeyType();
-
-        if (String.class == keyType) {
-            return key;
-        }
-
-        if (Integer.class == keyType) {
-            try {
-                return Integer.parseInt(key);
-            } catch (NumberFormatException ex) {
-                throw new InvalidRequestParamEx(key, "key", "String",
-                        "Keys for cache " + cacheType + " must be parsable into an integer");
-            }
-        }
-
-
-        throw new InvalidRequestParamEx(cacheType.name(), "cacheName", "String",
-                "Element evict is not currently supported for " + cacheType);
-    }
-
-
 }
