@@ -1,44 +1,31 @@
 package gov.nysenate.ess.core.dao.unit;
 
-import com.google.common.eventbus.EventBus;
-import gov.nysenate.ess.core.model.cache.ContentCache;
-import gov.nysenate.ess.core.model.personnel.ResponsibilityHead;
+import gov.nysenate.ess.core.model.cache.CacheType;
 import gov.nysenate.ess.core.model.unit.Location;
 import gov.nysenate.ess.core.model.unit.LocationId;
 import gov.nysenate.ess.core.service.base.CachingService;
-import gov.nysenate.ess.core.service.cache.EhCacheManageService;
-import net.sf.ehcache.Cache;
-import net.sf.ehcache.Element;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
-import java.util.Collection;
 import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.stream.Collectors;
 
 @Service
-public class LocationDao implements CachingService<String> {
-
+public class LocationDao extends CachingService<String, LocationDao.LocationCacheTree> {
     private static final Logger logger = LoggerFactory.getLogger(LocationDao.class);
     private static final String LOCATION_CACHE_KEY = "location";
+    private final SqlLocationDao sqlLocationDao;
 
-    private SqlLocationDao sqlLocationDao;
-    private EventBus eventBus;
-    private EhCacheManageService cacheManageService;
-    private volatile Cache locationCache;
-
-    public LocationDao(SqlLocationDao sqlLocationDao, EventBus eventBus, EhCacheManageService cacheManageService) {
+    @Autowired
+    public LocationDao(SqlLocationDao sqlLocationDao) {
         this.sqlLocationDao = sqlLocationDao;
-        this.eventBus = eventBus;
-        this.cacheManageService = cacheManageService;
-
-        this.eventBus.register(this);
-        this.locationCache = this.cacheManageService.registerEternalCache(getCacheType().name());
-        if (this.cacheManageService.isWarmOnStartup()) {
+        if (warmOnStartup) {
             cacheLocations();
         }
     }
@@ -56,7 +43,7 @@ public class LocationDao implements CachingService<String> {
      * @return
      */
     public List<Location> getLocations() {
-        return this.getLocations(false);
+        return getLocations(false);
     }
 
     /**
@@ -76,16 +63,6 @@ public class LocationDao implements CachingService<String> {
     }
 
     /**
-     * Get active locations that are managed by the given {@code responsibilityHead}.
-     * @return A list of matching locations or an empty list if no matches were found.
-     */
-    public List<Location> getLocationsByResponsibilityHeads(Collection<ResponsibilityHead> responsibilityHeads) {
-        return getLocations(true).stream()
-                .filter(loc -> responsibilityHeads.contains(loc.getResponsibilityHead()))
-                .collect(Collectors.toList());
-    }
-
-    /**
      * Searches for locations where code matches {@code term}.
      * @param term
      * @return A list of matching locations
@@ -94,9 +71,8 @@ public class LocationDao implements CachingService<String> {
         return sqlLocationDao.searchLocations(term);
     }
 
-    private static final class LocationCacheTree {
-
-        private HashMap<LocationId, Location> locationTreeMap = new HashMap<>();
+    static final class LocationCacheTree {
+        private final HashMap<LocationId, Location> locationTreeMap = new HashMap<>();
 
         public LocationCacheTree(List<Location> locations) {
             for (Location loc : locations) {
@@ -113,31 +89,12 @@ public class LocationDao implements CachingService<String> {
         }
     }
 
-    /** --- Caching Service Implemented Methods --- */
-
     /**
      * {@inheritDoc}
      */
     @Override
-    public ContentCache getCacheType() {
-        return ContentCache.LOCATION;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public void evictContent(String key) {
-        locationCache.remove(key);
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public void evictCache() {
-        logger.info("Clearing {} cache..", getCacheType());
-        locationCache.removeAll();
+    public CacheType cacheType() {
+        return CacheType.LOCATION;
     }
 
     /**
@@ -145,43 +102,21 @@ public class LocationDao implements CachingService<String> {
      */
 
     private LocationCacheTree getLocationCacheTree() {
-        Element element = getCachedLocationMap();
-        if (cacheIsEmpty(element)) {
+        LocationCacheTree tree = cache.get(LOCATION_CACHE_KEY);;
+        if (tree == null || tree.getLocations().isEmpty()) {
             cacheLocations();
-            element = getCachedLocationMap();
+            tree = cache.get(LOCATION_CACHE_KEY);;
         }
-        return (LocationCacheTree) element.getObjectValue();
+        return tree;
     }
 
-    private boolean cacheIsEmpty(Element element) {
-        if (element == null) {
-            return true;
-        }
-        LocationCacheTree locationCacheTree = (LocationCacheTree) element.getObjectValue();
-        return locationCacheTree.getLocations().size() == 0;
-    }
-
-    private Element getCachedLocationMap() {
-        Element element;
-        locationCache.acquireReadLockOnKey(LOCATION_CACHE_KEY);
-        try {
-            element = locationCache.get(LOCATION_CACHE_KEY);
-        } finally {
-            locationCache.releaseReadLockOnKey(LOCATION_CACHE_KEY);
-        }
-        return element;
+    @Override
+    protected Map<String, LocationCacheTree> initialEntries() {
+        return Map.of(LOCATION_CACHE_KEY, new LocationCacheTree(sqlLocationDao.getLocations(false)));
     }
 
     @Scheduled(cron = "${cache.cron.location:0 0 0 * * *}")
     private void cacheLocations() {
-        logger.info("Warming Locations cache...");
-        locationCache.acquireWriteLockOnKey(LOCATION_CACHE_KEY);
-        try {
-            locationCache.remove(LOCATION_CACHE_KEY);
-            locationCache.put(new Element(LOCATION_CACHE_KEY, new LocationCacheTree(sqlLocationDao.getLocations(false))));
-        } finally {
-            locationCache.releaseWriteLockOnKey(LOCATION_CACHE_KEY);
-        }
-        logger.info("Done caching locations.");
+        clearCache(true);
     }
 }
