@@ -1,75 +1,63 @@
 package gov.nysenate.ess.core.service.base;
 
-import com.google.common.eventbus.Subscribe;
-import gov.nysenate.ess.core.model.cache.CacheEvictEvent;
-import gov.nysenate.ess.core.model.cache.CacheEvictIdEvent;
-import gov.nysenate.ess.core.model.cache.CacheWarmEvent;
-import gov.nysenate.ess.core.model.cache.ContentCache;
+import com.google.common.eventbus.EventBus;
+import gov.nysenate.ess.core.model.cache.CacheType;
+import gov.nysenate.ess.core.service.cache.EssCacheManager;
+import org.ehcache.Cache;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
+
+import javax.annotation.PostConstruct;
+import java.util.Map;
 
 /**
- * Defines methods that are required to manage a service that depends on one or more caches
- * @param <ContentId>
+ * Defines methods that are required to manage a service that depends on one or more caches.
  */
-public interface CachingService<ContentId>
-{
+public abstract class CachingService<K, V> {
+    private static final Logger logger = LoggerFactory.getLogger(CachingService.class);
+    @Autowired
+    protected EventBus eventBus;
+    @Value("${cache.warm.onstartup:false}")
+    protected boolean warmOnStartup;
+    protected Cache<K, V> cache;
+
+    @PostConstruct
+    private void init() {
+        // TODO: correct size
+        this.cache = EssCacheManager.createCache(this, 100);
+    }
+
     /**
-     * @return A ContentCache enum specifying which cache this is
+     * @return Specifies which cache this is.
      */
-    ContentCache getCacheType();
+    public abstract CacheType cacheType();
+
+    // TODO: Make sure everything has some initial size
+    protected Map<K, V> initialEntries() {
+        return Map.of();
+    }
 
     /**
      * Evicts a single item from the cache based on the given content id
      */
-    void evictContent(ContentId contentId);
+    public void evictContent(K k) {
+        cache.remove(k);
+    }
+
+    public int expiryTimeSeconds() {
+        return -1;
+    }
 
     /**
      * Clears all the cache entries from cache.
      */
-    void evictCache();
-
-    /**
-     * Pre-fetch a subset of currently active data and store it in the cache.
-     * If no override is provided, just evict the cache
-     */
-    default void warmCache() {
-        evictCache();
-    }
-
-    /**
-     * If a CacheEvictEvent is sent out on the event bus, the caching service
-     * should check to see if it has any affected caches and clear them.
-     *
-     * @param evictEvent CacheEvictEvent
-     */
-    @Subscribe
-    default void handleCacheEvictEvent(CacheEvictEvent evictEvent) {
-        if (evictEvent.affects(getCacheType())) {
-            evictCache();
-        }
-    }
-
-    /**
-     * Intercept an evict Id event and evict the specified content
-     * if the caching service has any of the affected caches
-     * @param evictIdEvent CacheEvictIdEvent
-     */
-    @Subscribe
-    default void handleCacheEvictIdEvent(CacheEvictIdEvent<ContentId> evictIdEvent) {
-        if (evictIdEvent.affects(getCacheType())) {
-            evictContent(evictIdEvent.getContentId());
-        }
-    }
-
-    /**
-     * If a CacheWarmEvent is sent out on the event bus, the caching service
-     * should check to if it has any affected caches and warm them.
-     *
-     * @param warmEvent CacheWarmEvent
-     */
-    @Subscribe
-    default void handleCacheWarmEvent(CacheWarmEvent warmEvent) {
-        if (warmEvent.affects(getCacheType())) {
-            warmCache();
+    public void clearCache(boolean warmCache) {
+        logger.info("Clearing " + cacheType().name() + " cache...");
+        cache.clear();
+        if (warmCache) {
+            cache.putAll(initialEntries());
         }
     }
 }
