@@ -6,23 +6,21 @@ import gov.nysenate.ess.core.dao.pec.assignment.PersonnelTaskAssignmentDao;
 import gov.nysenate.ess.core.dao.pec.task.PersonnelTaskDao;
 import gov.nysenate.ess.core.dao.pec.task.detail.PersonnelTaskDetailDao;
 import gov.nysenate.ess.core.dao.personnel.EmployeeDao;
-import gov.nysenate.ess.core.model.cache.ContentCache;
+import gov.nysenate.ess.core.model.cache.CacheType;
 import gov.nysenate.ess.core.model.pec.PersonnelTask;
 import gov.nysenate.ess.core.model.pec.PersonnelTaskAssignment;
 import gov.nysenate.ess.core.model.pec.PersonnelTaskType;
 import gov.nysenate.ess.core.model.pec.video.PersonnelTaskAssignmentGroup;
 import gov.nysenate.ess.core.model.personnel.Employee;
 import gov.nysenate.ess.core.service.base.CachingService;
-import gov.nysenate.ess.core.service.cache.EhCacheManageService;
-import net.sf.ehcache.Cache;
-import net.sf.ehcache.Element;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.time.Duration;
 import java.util.HashSet;
 import java.util.List;
+import java.util.Map;
 import java.util.Set;
 import java.util.function.Function;
 
@@ -31,33 +29,45 @@ import java.util.function.Function;
  * and utilizing caching to improve performance.
  */
 @Service
-public class CachedPersonnelTaskService implements PersonnelTaskService, CachingService<String> {
+public class CachedPersonnelTaskService
+        extends CachingService<String, ImmutableSortedMap<Integer, PersonnelTask>>
+        implements PersonnelTaskService {
 
     private static final Logger logger = LoggerFactory.getLogger(CachedPersonnelTaskService.class);
 
     /** Key used to store a map of all tasks */
     private static final String TASK_MAP_KEY = "TASK_MAP_KEY";
     /** Number of seconds the task map is valid in cache before it is evicted */
-    private static final Long cacheEvictTime = Duration.ofMinutes(5).getSeconds();
+    private static final int cacheEvictTime = 300;
 
     private final PersonnelTaskDao taskDao;
     private final ImmutableMap<PersonnelTaskType, PersonnelTaskDetailDao<?>> taskDetailDaoMap;
-    private final Cache taskCache;
     private final EmployeeDao employeeDao;
     private final PersonnelTaskAssignmentDao personnelTaskAssignmentDao;
 
+    @Autowired
     public CachedPersonnelTaskService(PersonnelTaskDao taskDao,
                                       List<PersonnelTaskDetailDao<?>> taskDetailDaos,
-                                      EhCacheManageService cacheManageService,
-                                      EventBus eventBus,
                                       EmployeeDao employeeDao,
                                       PersonnelTaskAssignmentDao personnelTaskAssignmentDao) {
         this.taskDao = taskDao;
         this.taskDetailDaoMap = Maps.uniqueIndex(taskDetailDaos, PersonnelTaskDetailDao::taskType);
-        this.taskCache = cacheManageService.registerTimeBasedCache(getCacheType().name(), cacheEvictTime);
         this.employeeDao = employeeDao;
         this.personnelTaskAssignmentDao = personnelTaskAssignmentDao;
-        eventBus.register(this);
+    }
+
+    @Override
+    public int expiryTimeSeconds() {
+        return cacheEvictTime;
+    }
+
+    @Override
+    protected Map<String, ImmutableSortedMap<Integer, PersonnelTask>> initialEntries() {
+        ImmutableSortedMap<Integer, PersonnelTask> map = taskDao.getAllTasks().stream()
+                .map(task -> taskDetailDaoMap.get(task.getTaskType()).getTaskDetails(task))
+                .collect(ImmutableSortedMap.toImmutableSortedMap(
+                        Comparable::compareTo, PersonnelTask::getTaskId, Function.identity()));
+        return Map.of(TASK_MAP_KEY, map);
     }
 
     /* --- PersonnelTaskService implementations --- */
@@ -81,41 +91,26 @@ public class CachedPersonnelTaskService implements PersonnelTaskService, Caching
     @Override
     public List<PersonnelTask> getPersonnelTasks(boolean activeOnly) {
         return getTaskMap().values().stream()
-                .filter(task -> !activeOnly || task.isActive())
-                .collect(ImmutableList.toImmutableList());
+                .filter(task -> !activeOnly || task.isActive()).toList();
     }
 
     @Override
     public List<PersonnelTask> getActiveTasksInGroup(PersonnelTaskAssignmentGroup assignmentGroup) {
         return getPersonnelTasks(true).stream()
-                .filter(task -> task.getAssignmentGroup() == assignmentGroup)
-                .collect(ImmutableList.toImmutableList());
+                .filter(task -> task.getAssignmentGroup() == assignmentGroup).toList();
     }
 
     /* --- CachingService implementations --- */
 
     @Override
-    public ContentCache getCacheType() {
-        return ContentCache.PERSONNEL_TASK;
+    public CacheType cacheType() {
+        return CacheType.PERSONNEL_TASK;
     }
 
     @Override
     public void evictContent(String key) {
         // Removing all since there is only one key
-        evictCache();
-    }
-
-    @Override
-    public void evictCache() {
-        logger.info("Clearing personnel task cache...");
-        taskCache.removeAll();
-    }
-
-    @Override
-    public void warmCache() {
-        evictCache();
-        logger.info("Warming personnel task cache...");
-        getPersonnelTasks(false);
+        super.clearCache(false);
     }
 
     public void markTasksComplete() {
@@ -133,10 +128,8 @@ public class CachedPersonnelTaskService implements PersonnelTaskService, Caching
         }
 
         for (Employee employee : employeesToMarkComplete) {
-
             List<PersonnelTaskAssignment> assignments =
                     personnelTaskAssignmentDao.getAssignmentsForEmp(employee.getEmployeeId());
-
             for (PersonnelTaskAssignment assignment : assignments) {
                 if (!assignment.isCompleted()) {
                     personnelTaskAssignmentDao.setTaskComplete(
@@ -149,33 +142,11 @@ public class CachedPersonnelTaskService implements PersonnelTaskService, Caching
 
     /* --- Internal Methods --- */
 
-    @SuppressWarnings("unchecked")
     private ImmutableSortedMap<Integer, PersonnelTask> getTaskMap() {
-        taskCache.acquireReadLockOnKey(TASK_MAP_KEY);
-        Element result = taskCache.get(TASK_MAP_KEY);
-        taskCache.releaseReadLockOnKey(TASK_MAP_KEY);
-        final ImmutableSortedMap<Integer, PersonnelTask> taskMap;
-        if (result == null) {
-            taskMap = loadTaskMap();
-            taskCache.acquireWriteLockOnKey(TASK_MAP_KEY);
-            taskCache.put(new Element(TASK_MAP_KEY, taskMap));
-            taskCache.releaseWriteLockOnKey(TASK_MAP_KEY);
-        } else {
-            taskMap = (ImmutableSortedMap<Integer, PersonnelTask>) result.getObjectValue();
+        var value = cache.get(TASK_MAP_KEY);
+        if (value == null) {
+            clearCache(true);
         }
-        return taskMap;
-    }
-
-    private ImmutableSortedMap<Integer, PersonnelTask> loadTaskMap() {
-        return taskDao.getAllTasks().stream()
-                .map(this::getDetailedTask)
-                .collect(ImmutableSortedMap.toImmutableSortedMap(
-                        Comparable::compareTo, PersonnelTask::getTaskId, Function.identity())
-                );
-    }
-
-    private PersonnelTask getDetailedTask(PersonnelTask basicTask) {
-        return taskDetailDaoMap.get(basicTask.getTaskType())
-                .getTaskDetails(basicTask);
+        return value;
     }
 }
